SINBA.saveRData

SINBA.saveRData is a function to save complicated list object generated by certain steps of SINBA pipeline (e.g. load network files, 'pre-load', load expression files, 'exp-load', calculate driver pairs activity as 'combo-AC', predict drug pairs table 'ms-tab').This function is not essential, but it is highly suggested for easier pipeline step checkout and reference.

Assigning the step name to save the RData for easier reference. Calling SINBA.loadRData to load the corresponding step RData, users can avoid repeating the former steps.

Usage

SINBA.saveRData(SINBA.par = NULL, step = "pre-load")

Arguments

SINBA.par

list, stores all related datasets from driver selection and driver combination prediction pipeline step.

step

character, name of the pipeline step decided by user for easier reference.

Examples

if (FALSE) { # \dontrun{
SINBA.par <- list()
SINBA.par$out.dir.DATA <- system.file('demo1','/DATA/',package = "SINBA")
SINBA.loadRData(SINBA.par=SINBA.par,step='pre-load')
SINBA.saveRData(SINBA.par=SINBA.par,step='pre-load')
} # }