cal.SINBA.Activity

This function will Calculate Activity Value for Each Driver or seed-partner pairs.

Usage

cal.SINBA.Activity(
  SINBA_originalID = NULL,
  target_list = NULL,
  cal_mat = NULL,
  es.method = "weightedmean",
  std = TRUE
)

Arguments

SINBA_originalID

character vector or the return data frame generated by combineNet2target_list with return_type="SINBA_originalID". SINBA_originalID contains the seed-partner driver pairs; seed gene name and partner gene name are seperated by colon. SINBA_originalID is optional.

target_list

list, the driver-to-target list object.target_list is generated by get.combined.network and contains all the driver gene regulons.

cal_mat

numeric matrix, the expression matrix of genes/transcripts.

es.method

character, method applied to calculate the activity value. User can choose from c ("mean", "weightedmean", "maxmean", "absmean", "max" and "min"). Default is "weightedmean".

std

logical, if TRUE, the expression matrix will be normalized by column. Default is TRUE.

Examples

sinba_originalID<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN","STAT3","PTCRA"),return_type="SINBA_originalID")
#> Error: object 'SINBA.par' not found
cal_mat<-Biobase::exprs(SINBA.par$cal.eset)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'exprs': object 'SINBA.par' not found
sinba_ac_mat<-cal.SINBA.Activity(SINBA_originalID=sinba_originalID,target_list=SINBA.par$combined.network$target_list,cal_mat=cal_mat)
#> Error: object 'cal_mat' not found
single_ac_mat<-cal.SINBA.Activity(target_list=SINBA.par$combined.network$target_list,cal_mat=cal_mat)
#> Error: object 'cal_mat' not found
if (FALSE) { # \dontrun{
} # }