convert_deltaZ2_matrix

Convert a dataframe with selected columns(including the deltaZ score column) to a matrix.

Usage

convert_deltaZ2_matrix(
  df,
  use_row.col,
  use_column.col,
  deltaZ.col,
  use_method = "max"
)

Arguments

df

a data frame with deltaZ scores, it could be created by run generate.masterTable.SINBA .

use_row.col

column name in the df to be used as rows in the output matrix.

use_column.col

colum name in the df to be used as columns in the output matrix.

deltaZ.col

column name with the deltaZ scores, the values will be used in the output matrix.

use_method

the method used to aggregate the deltaZ values from same seed/partner drivers. Options: "max", "mean", "min" and "median". default="max"

Examples

ms_tab<-SINBA.par$ms_tab
#> Error: object 'SINBA.par' not found
df<-convert_deltaZ2_matrix(df=ms_tab,use_row.col="SINBA.seed_geneSymbol",use_column.col="SINBA.partner_geneSymbol",deltaZ.col="delta_Z.G3.Vs.others_SINBA",use_method="max")
#> Error: object 'ms_tab' not found
if (FALSE) { # \dontrun{
} # }