draw.NetBID_SINBA

GSEA (gene set enrichment analysis) plot for the Synergy Inference by data-driven Network-based Bayesian Analysis (SINBA) analysis results.

draw.GSEA.NetBID.SINBA will generate a GSEA plot for Synergy Inference by data-driven Network-based Bayesian Analysis (SINBA) analysis results. SINBA calculates the synergistic effect between a seed driver and a partner driver. The plot includes the GSEA plot for the seed driver and the partner driver independently and the GSEA plot for the combination for the seed driver to each partner driver. The statistics on the plot include the differentiated expression (DE), differentiated activity (DA) for each driver, and the different Z (deltaZ) showing the difference between the combination of the seed and the partner driver to the sum of the original Z statistics.

This is a plot function to draw GSEA for synergistic effect prediction between the seed driver and a list of partner drivers. User need to input the differentiated expression information, and choose to display the target genes in one row or two rows, by selecting black color or red to blue color bar.

Usage

draw.NetBID_SINBA(
  DE = NULL,
  name_col = NULL,
  profile_col = NULL,
  profile_trend = "pos2neg",
  seed_driver = NULL,
  partner_driver_list = NULL,
  seed_driver_label = NULL,
  partner_driver_label = NULL,
  driver_DA_Z = NULL,
  driver_DE_Z = NULL,
  target_list = NULL,
  target_list_merge = NULL,
  DA_Z_merge = NULL,
  diff_Z = NULL,
  top_driver_number = 10,
  top_order = "merge",
  target_nrow = 2,
  target_col = "RdBu",
  target_col_type = "PN",
  left_annotation = "",
  right_annotation = "",
  main = "",
  profile_sig_thre = 0,
  Z_sig_thre = 1.64,
  pdf_file = NULL
)

Arguments

DE

data.frame,the differentiated expression results. This data.frame could be generated by using getDE.limma.2G or getDE.BID.2G. If user want to generate this data.frame by other strategies, the rownames must be the gene names or need one column to be the gene name (set in name_col) and must contain the columns indicating the differentiated expression profile.

name_col

character, the name of the column in DE, which contains the gene name. If NULL, will use the rownames of DE. Default is NULL.

profile_col

character, the name of the column in DE, which will be used as the differentiated expression profile. If DE is created by getDE.limma.2G or getDE.BID.2G, this parameter could be 'logFC' or 't'.

profile_trend

character, the choice of how to display the profile, from high/positive to low/negative ('pos2neg') or low/negative to high/positive ('neg2pos').Default is 'pos2neg'.

seed_driver

character, name for the seed driver.

partner_driver_list

a vector of characters, name for the partner driver list.

seed_driver_label

character, label for the seed driver displayed on the plot. Default is seed_driver.

partner_driver_label

a vector of characters, label for the partner driver list displayed on the plot. Default is partner_driver_list

driver_DA_Z

a vector of numeric values, the Z statistics of differentiated activity (DA) for the driver list. Better to give name to the vector, otherwise will automatically use driver list (seed + partner) as the name.

driver_DE_Z

a vector of numeric values, the Z statistics of differentiated expression (DE) for the driver list. Better to give name to the vector, otherwise will automatically use driver list (seed + partner) as the name.

target_list

a list for the target gene information for the drivers. The names for the list must contain the driver in driver list (seed + partner) Each object in the list must be a data.frame and should contain one column ("target") to save the target genes. Strongly suggest to follow the NetBID2 pipeline, and the target_list could be automatically generated by get_net2target_list by running get.SJAracne.network.

target_list_merge

a list for the target gene information for thecombination of the seed driver to partner drivers saperately. The names for the list must contain the driver in partner_driver_list Each object in the list must be a data.frame and should contain one column ("target") to save the target genes. Strongly suggest to follow the NetBID2 pipeline, and the target_list_merge could be automatically generated by merge_target_list.

DA_Z_merge

a vector of numeric values, the Z statistics of differentiated activity (DA) for the combination of the seed driver to partner drivers saperately. Better to give name to the vector, otherwise will automatically use partner driver list as the name.

top_driver_number

numeric, number for the top significant partner drivers to be displayed on the plot. Default is 10.

top_order

character, choice of order pattern used to display the partner drivers. Two options,'merge' or 'diff'. 'merge' means the partner drivers will be sorted by the combined Z statistics. 'diff' means the partner drivers will be sorted by the delta Z statistics. Default is 'merge'.

target_nrow

numeric, number of rows for each driver display on the plot. Two options, 1 or 2. If set to 1, the target genes' position on the profile will be displayed in one row. If set to 2, the target genes' position on the profile will be displayed in two rows, with positive regulated genes displayed on the first row and negative regulated genes displayed on the second row. Default is 2.

target_col

character, choice of color pattern used to display the targets. Two options,'black' or 'RdBu'. If set to 'black', the lines will be colored in black. If set to 'RdBu', the lines will be colored into Red to Blue color bar. If target_col_type is set as 'PN', the positive regulated genes will be colored in red and negative regulated genes in blue. If target_col_type is set as 'DE', the color for the target genes is set according to its value in the differentiated expression profile, with significant high set for red and low for blue. The significant threshold is set by profile_sig_thre. Default is 'RdBu'.

target_col_type

character, choice of the pattern used to display the color for target genes, only work when target_col is set as 'RdBu'. Two options,'PN' or 'DE'. If set as 'PN', the positive regulated genes will be colored in red and negative regulated genes in blue. If set as 'DE', the color for the target genes is set according to its value in the differentiated expression profile, Default is 'PN'.

left_annotation

character, annotation displayed on the left of the figure representing left condition of the rank_profile. Default is "".

right_annotation

character, annotation displayed on the right of the figure representing right condition of the rank_profile. Default is "".

main

character, title for the plot. Default is "".

profile_sig_thre

numeric, threshold for the absolute values in profile to be treated as significance. Target genes without signifcant values in the profile will be colored in grey. Only work when target_col_type is set as "DE" and target_col is set as "RdBu". Default is 0.

Z_sig_thre

numeric, threshold for the Z statistics in driver_DA_Z and driver_DE_Z to be treated as signifcance. Only signifcant values will have background color. Default is 1.64.

pdf_file

character, file path for the pdf file to save the figure into pdf format.If NULL, will not generate pdf file. Default is NULL.

Value

logical value indicating whether the plot has been successfully generated

Examples