Preload drug database files into R workspace for SINBA

drugdb.preload is a pre-processing function for SINBA. It preloads needed drug databases into R workspace,and saves it locally under data/ directory.

Usage

drugdb.preload(db = NULL, main.dir = NULL, out.dir = NULL)

Arguments

db

A vector of characters, users can choose from "BPdb","ChEMBL","cMAP","Drugbank_V5.1", "rDGIdb_1.8.0" and "Repurposing_HUB". If NULL, all the drug databases will be used. Default is NULL.

main.dir

character, the main directory for SINBA package. If NULL, will be system.file(package = "SINBA"). Default is NULL.

out.dir

charater, the directory to save drug database information. If NULL, will be system.file(package = "SINBA"). Default is NULL.

Details

Users need to choose the drug databse sources(e.g. "BPdb","ChEMBL","cMAP","Drugbank_V5.1", "rDGIdb_1.8.0","Repurposing_HUB")

Examples

drugdb.preload()
#> db not set, will use all the available databases: BPdb,ChEMBL,cMAP,Drugbank_V5.1,rDGIdb_1.8.0,Repurposing_HUB!
#> main.dir not set, will use package directory: /private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data
#> out.dir not set, will use package directory: /private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data
#> Warning: cannot open compressed file '/private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data/Drugdb_SINBA.rda', probable reason 'No such file or directory'
#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

if (FALSE) { # \dontrun{
drugdb.preload()
} # }