generate.masterTable.SINBA

The master table gathers all the DE (differential expression analysis), DA (differential activity analysis of of single driver) and DA.SINBA (differential activity analysis of driver pairs) from multiple comparisons.

Usage

generate.masterTable.SINBA(
  use_comp = NULL,
  DA.SINBA = NULL,
  DA = NULL,
  DE = NULL,
  ftest = NULL,
  z_col = "Z-statistics",
  display_col = c("logFC", "P.Value"),
  deltaZ_method = "max"
)

Arguments

use_comp

a vector of characters, the name of multiple comparisons. It will be used to name the columns of master table.

DA.SINBA

list, a list of DA comparisons of driver pairs activity, each comparison is a data.frame. The element name in the list must contain the name in use_comp.

DA

list, a list of DA comparisons of single driver activity, each comparison is a data.frame. The element name in the list must contain the name in use_comp.

DE

list, a list of DE comparisons, each comparison is a data.frame. The element name in the list must contain the name in use_comp.

ftest

data.frame, the output of seed2partners_ftest, summarized the fisher exact test resutls of seed and partern drivers.

z_col

character, name of the column in DE and DA contains the Z statistics. Default is "Z-statistics".

display_col

character, name of the column in DE and DA need to be kept in the master table (e.g."P.Value","adj.P.Val" and "logFC"). Default is c("logFC","P.Value").

deltaZ_method

character, name of the method to be used for calculate deltaZ value. Users can choose from "max" and "mean". Default value is "max".

Examples

use_comp<-names(SINBA.par$DA)
#> Error: object 'SINBA.par' not found
out<-generate.masterTable.SINBA(use_comp=use_comp,DA.SINBA=SINBA.par$SINBA.DA,DA=SINBA.par$DA,DE=SINBA.par$DE,ftest=SINBA.par$SINBA.ftest,z_col='Z-statistics',display_col=c('logFC','P.Value'),deltaZ_method="max")
#> Error: object 'use_comp' not found
if (FALSE) { # \dontrun{
} # }