get.SINBA.DE.2G

Differential Expression Analysis and Differential Activity Analysis Between 2 Sample Groups. get.SINBA.DE.2G is a function performs differential gene expression analysis and differential driver activity analysis between control group (parameter G0) and experimental group (parameter G1), using different methods.

Usage

get.SINBA.DE.2G(
  eset = NULL,
  G1 = NULL,
  G0 = NULL,
  G1_name = NULL,
  G0_name = NULL,
  verbose = TRUE,
  do.log2 = F,
  test.use = "limma",
  random_effect = NULL
)

Arguments

eset

ExpressionSet class object, contains gene expression data or driver activity data.

G1

a vector of characters, the sample names of experimental group.

G0

a vecotr of characters, the sample names of control group.

G1_name

character, the name of experimental group (e.g. "Male"). Default is "G1".

G0_name

character, the name of control group (e.g. "Female"). Default is "G0".

verbose

logical, if TRUE, sample names of both groups will be printed. Default is TRUE.

test.use

character, users can choose one of the methods c("limma","t.test","wilcox"). Default is "limma"

random_effect

a vector of characters, vector or factor specifying a blocking variable.Default is NULL, no random effect will be considered.

Value

Return a data frame. Rows are genes/drivers, columns are "ID","Z-statistics","P.Value","adj.P.Val","logFC"

Examples

phe_info <- Biobase::pData(SINBA.par$cal.eset)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'SINBA.par' not found
phe_info$subgroup<-phe_info$das_sensitivity_category1
#> Error: object 'phe_info' not found
G0  <- rownames(phe_info)[which(phe_info$`subgroup`=="resistant")] # get sample list for G0
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rownames': object 'phe_info' not found
G1  <- rownames(phe_info)[which(phe_info$`subgroup`=="sensitive")] # get sample list for G1
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rownames': object 'phe_info' not found
ac_eset<-generate.SINBA.eset(exp_mat=ac_mat, phenotype_info=phe_info, feature_info=NULL, annotation_info="")
#> Error: object 'ac_mat' not found
DE_gene_limma <- get.SINBA.DE.2G(eset=SINBA.par$cal.eset,G1=G1,G0=G0,G1_name="sensitive",G0_name='resistant',test.use="limma")
#> Error: object 'SINBA.par' not found
DA_gene_limma <- get.SINBA.DE.2G(eset=ac_eset,G1=G1,G0=G0,G1_name="sensitive",G0_name='resistant',test.use="limma")
#> Error: object 'ac_eset' not found
if (FALSE) { # \dontrun{
} # }