get.single.network

Read SJARACNe Network Result and Return it as List Object

get.single.network reads SJARACNe network construction result and returns a list object including 1.network data frame, 2.driver-to-target list and 3.igraph object wrapped inside.

In the demo, "consensus_network_ncol_.txt" file will be read and convert into a list object. This list contains three elements, network_data, target_list and igraph_obj. network_dat is a data.frame, contains all the information of the network SJARACNe constructed. target_list is a driver-to-target list object. Please check details in get_net2target_list. igraph_obj is an igraph object used to save this directed and weighted network. Each edge of the network has two attributes, weight and sign. weight is the "MI (mutual information)" value and sign is the sign of the spearman correlation coefficient (1, positive regulation; -1, negative regulation).

Usage

get.single.network(network_file = NULL)

Arguments

network_file

character, the path for storing network file. For the output of SJAracne, the name of the network file will be "consensus_network_ncol_.txt" under the output directory.

Value

Return a list containing three elements, network_dat, target_list and igraph_obj.

Examples

out_net_file<-"path_to_network_file/consensus_network_ncol_.txt"
network  <- get.single.network(network_file=out_net_file)
#> Warning: cannot open file 'path_to_network_file/consensus_network_ncol_.txt': No such file or directory
#> Error in file(file, "rt"): cannot open the connection

if (FALSE) { # \dontrun{
} # }