l2l_ftest

l2l_ftest

Usage

l2l_ftest(gsc1, gsc2, N)

Arguments

gsc1

a list of genesets, each geneset contains multiple gene names.

gsc2

a list of genesets, each geneset contains multiple gene names.

N

mumeric, total unique gene number in gsc1 and gsc2.

Value

Return a table of all fisher exact testing results.