queryDGIdb.SINBA

a wrapper to query DGIdb API with gene name. Deprecated Use queryDgidb_V5_gene() instead.

Usage

queryDGIdb.SINBA(
  genes = NULL,
  source_Databases = c("CGI", "CIViC", "COSMIC", "CancerCommons", "ChemblInteractions",
    "ClearityFoundationBiomarkers", "ClearityFoundationClinicalTrial", "DTC", "DoCM",
    "DrugBank", "FDA", "GuideToPharmacology", "JAX-CKB", "MyCancerGenome",
    "MyCancerGenomeClinicalTrial", "NCI", "OncoKB", "PharmGKB", "TALC", "TEND", "TTD",
    "TdgClinicalTrial"),
  gene_Categories = c("ABC TRANSPORTER", "B30_2 SPRY DOMAIN", "CELL SURFACE",
    "CLINICALLY ACTIONABLE", "CYTOCHROME P450", "DNA DIRECTED RNA POLYMERASE",
    "DNA REPAIR", "DRUG METABOLISM", "DRUG RESISTANCE", "DRUGGABLE GENOME", "ENZYME",
    "EXCHANGER", "EXTERNAL SIDE OF PLASMA MEMBRANE", "FIBRINOGEN",
    "G PROTEIN COUPLED RECEPTOR", "GROWTH FACTOR", "HISTONE MODIFICATION",
    "HORMONE ACTIVITY", "ION CHANNEL", "KINASE", "LIPASE", "LIPID KINASE",
    "METHYL TRANSFERASE", "MYOTUBULARIN RELATED PROTEIN PHOSPHATASE",
    "NEUTRAL ZINC METALLOPEPTIDASE", 
     "NUCLEAR HORMONE RECEPTOR",
    "PHOSPHATIDYLINOSITOL 3 KINASE", "PHOSPHOLIPASE", "PROTEASE", "PROTEASE INHIBITOR",
    "PROTEIN PHOSPHATASE", "PTEN FAMILY", "RNA DIRECTED DNA POLYMERASE",
    "SERINE THREONINE KINASE", "SHORT CHAIN DEHYDROGENASE REDUCTASE", "THIOREDOXIN",
    "TRANSCRIPTION FACTOR", "TRANSCRIPTION FACTOR BINDING",
    "TRANSCRIPTION FACTOR COMPLEX", "TRANSPORTER", "TUMOR SUPPRESSOR", "TYROSINE KINASE",
    "UNKNOWN"),
  interaction_Types = c("activator", "adduct", "agonist", "allosteric modulator",
    "antagonist", "antibody", "antisense oligonucleotide", "binder", "blocker",
    "chaperone", "cleavage", "cofactor", "inducer", "inhibitor",
    "inhibitory allosteric modulator", "inverse agonist", "ligand", "modulator",
    "multitarget", "n/a", "negative modulator", "other/unknown", "partial agonist",
    "partial antagonist", "positive modulator", "potentiator", "product of",
    "stimulator", "substrate", "suppressor", "vaccine"),
  return_raw = F
)

Arguments

genes

a character vector of genes for which drug interactions are queried.

source_Databases

a character vector of source databses to be used. Users can call helper function sourceDatabases() to find all the available database. To query all available databases, skip argument (highly recommend).

gene_Categories

a character vector of gene categories to be used. Users can call helper function geneCategories() to find all the available categories. To query all available categories, skip argument.

interaction_Types

a character vector of gene drug interaction types to be used. Users can call helper function interactionTypes() to find all the available interactionTypes. To query all available interaction types, skip argument.

return_raw

a logical value, default is FALSE, only return summarized information. If users want to check raw query data, please change to TRUE.

Value

a list contains 1.a data frame("summary_result") of summarized gene,drug and interaction information; 2.raw query data("raw_result") from DGIdb (optional).

Examples

genes<-c("ATM","BCL2L1","LCK","CDK4")
res<-queryDGIdb.SINBA(genes=genes,interaction_Types=c("inhibitor","antagonist"))
#> Error in queryDGIdb.SINBA(genes = genes, interaction_Types = c("inhibitor",     "antagonist")): could not find function "queryDGIdb.SINBA"
res<-queryDGIdb.SINBA(genes=genes,source_Databases =c("ChemblInteractions","TTD","OncoKB","MyCancerGenome","DrugBank","FDA"),interaction_Types=c("inhibitor","antagonist"))
#> Error in queryDGIdb.SINBA(genes = genes, source_Databases = c("ChemblInteractions",     "TTD", "OncoKB", "MyCancerGenome", "DrugBank", "FDA"), interaction_Types = c("inhibitor",     "antagonist")): could not find function "queryDGIdb.SINBA"