selectSP_driver_byDB

selectSP_driver_byDB is used to select driver genes based on their interaction drug information, including predicted blood-brain barrier penetration and drug developing phase.

Usage

selectSP_driver_byDB(
  genes = NULL,
  select_type = NULL,
  database = NULL,
  select_phase = NULL
)

Arguments

genes

a character vector of genes for which the targeting drug phase or blood brain barrier penetration information are queried.

select_type

character, which drug information is required. User can choose from "BP_prediction" and "drug_phase".

database

character, which database will be used. "BPdb" contains drug blood brain barrier penetrantion information, others contain drug developing phase information. User can choose from "BPdb", "ChEMBL","Repurposing_HUB" and "Drugbank_V5.1".

select_phase

a character vector of drug phases will be used. For checking detailed phase information in each database, try to call helper functions: ChEMBL_drugphase_levels(); Repurposing_HUB_drugphase_levels(); Drugbank_drugphase_levels().

Examples

drugdb.preload()
#> db not set, will use all the available databases: BPdb,ChEMBL,cMAP,Drugbank_V5.1,rDGIdb_1.8.0,Repurposing_HUB!
#> main.dir not set, will use package directory: /private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data
#> out.dir not set, will use package directory: /private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data
#> Warning: cannot open compressed file '/private/var/folders/ks/by4rh1y10ssbd07q9z4ycwq40000gp/T/RtmpioPFV1/temp_libpath2365270c4416/SINBA/data/Drugdb_SINBA.rda', probable reason 'No such file or directory'
#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection
genes<-c("MDM2","CDK4","CDK6","ATM","ATR","BCL2","BCL2L1","HDAC6","STAT3","JAK2")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="BP_prediction",database="BPdb")
#> Error in is.null(genes) || length(genes) == 0 || !is.character(genes) ||     genes == "": 'length = 10' in coercion to 'logical(1)'
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Drugbank_V5.1")
#> Error in is.null(genes) || length(genes) == 0 || !is.character(genes) ||     genes == "": 'length = 10' in coercion to 'logical(1)'
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Drugbank_V5.1",select_phase="approved")
#> Error in is.null(genes) || length(genes) == 0 || !is.character(genes) ||     genes == "": 'length = 10' in coercion to 'logical(1)'
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="ChEMBL")
#> Error in is.null(genes) || length(genes) == 0 || !is.character(genes) ||     genes == "": 'length = 10' in coercion to 'logical(1)'
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Repurposing_HUB")
#> Error in is.null(genes) || length(genes) == 0 || !is.character(genes) ||     genes == "": 'length = 10' in coercion to 'logical(1)'
if (FALSE) { # \dontrun{
} # }